Python/Flask/1/requirements.txt
# Core Flask
Flask
Werkzeug==2.3.8
Jinja2
itsdangerous
click
numpy==1.22.0
pandas
blinker

# Production WSGI Server
gunicorn

# Database / ORM
SQLAlchemy
Flask-SQLAlchemy
zipp==3.19.1
Flask-Migrate
alembic
psycopg2

# Authentication / Forms
Flask-Login
Flask-WTF
WTForms

# Security
bcrypt
passlib
cryptography==45.0.7
Flask-Talisman

# Session Management
Flask-Session
redis

# Rate Limiting
Flask-Limiter

# Caching
Flask-Caching==2.3.1
cachelib

# Environment Configuration
python-dotenv

# Timezone Support
tzdata

# Logging
loguru
